TAJO-1950

+++ tajo-catalog/tajo-catalog-common/src/main/java/org/apache/tajo/catalog/statistics/Histogram.java
@@ -30,7 +30,7 @@ public abstract class Histogram<KEY, BUCKET> {
    * @param key
    * @param change
    */
-  public abstract void updateBucket(KEY key, long change);
+  public abstract void updateBucket(KEY key, double change);
 
   /**
    *
